First version of filter_fq.pl
I started writing the program because we need to filter a paired end data set which was quality filtered.
Therefore one of the reads may fail the length filter, but the other might pass the test. This will result
in non-synchronised fq files which is a lot of work to fix.

To avoid this resyncronize job, both files are read at the same time and a complete block will be printed if
both reads pass the test.